requirements.txt
cffi==1.15.1
Pillow==9.3.0
pillow-heif==0.8.0
pycparser==2.21
